% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HemaScribe.R
\name{broad_classify}
\alias{broad_classify}
\title{Broad classifier}
\usage{
broad_classify(input, return.full = FALSE)
}
\arguments{
\item{input}{A SingleCellExperiment or Seurat object}

\item{return.full}{Whether to return full predictions directly (default: FALSE)}
}
\value{
An object of the same type as input with added annotations, or a dataframe with classifier results
}
\description{
This performs a broad-level annotation of bone marrow cell types using
bulk RNA-sequencing data curated from Haeomopedia and Immgen.
}
