% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download reference data}
\usage{
download_data(file, repo, dataDir = NA)
}
\arguments{
\item{file}{Name of file to be downloaded}

\item{repo}{GitHub repository from which file is downloaded}

\item{dataDir}{Path where the data should be downloaded (default: package-managed user directory)}
}
\description{
Downloads the reference data from the Github repo.  The user is prompted for
download when the package is first loaded, but this function can also be
called in a later session to update the reference data file.
}
