% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_mapping_data}
\alias{download_mapping_data}
\title{Download mapping reference data}
\usage{
download_mapping_data(dataDir = NA)
}
\arguments{
\item{dataDir}{Specifies where the mapping reference data should be downloaded. Should essentially never be altered by the user directly since the package looks for the data in a specific location.}
}
\description{
Downloads the mapping reference data from the Github repo. The user is prompted for
download when the package is first loaded, but this function can also be
called in a later session to update the mapping reference data file.
}
