% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HemaScribe.R
\name{gene_id_to_symbol}
\alias{gene_id_to_symbol}
\title{Translate Ensembl IDs to gene names}
\usage{
gene_id_to_symbol(x)
}
\arguments{
\item{x}{List of Ensembl IDs}
}
\value{
List of gene names
}
\description{
Translate Ensembl IDs to gene names
}
