% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HemaScribe.R
\name{gene_symbol_to_id}
\alias{gene_symbol_to_id}
\title{Translate gene names to Ensembl IDs}
\usage{
gene_symbol_to_id(x)
}
\arguments{
\item{x}{List of gene names}
}
\value{
List of Ensembl IDs
}
\description{
Translate gene names to Ensembl IDs
}
