% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HemaScribe.R
\name{HemaScribe}
\alias{HemaScribe}
\title{Two-stage classifier}
\usage{
HemaScribe(input, prefilter = 0, reference = "WT", return.full = FALSE)
}
\arguments{
\item{input}{A SingleCellExperiment or Seurat object}

\item{prefilter}{Hematopoietic score threshold below which cells shall be excluded.  (default: 0)}

\item{reference}{Either "WT" or "5FU" (default: "WT")}

\item{return.full}{Whether to return full predictions directly (default: FALSE)}
}
\value{
An object of the same type as input with added annotations, or dataframes with classifier results.
}
\description{
This first performs the broad annotation of bone marrow cell types and then
applies the fine annotation to the cells that are annotated as HSPC.
Combined, HSPC-specific, and GMP-specific annotations are extracted from
these predictions.  An optional prefiltering step based on hematopoietic gene
scores is available; if no threshold is set, then the scores are merely
reported.
}
