% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StromaScribe.R
\name{stroma_classify}
\alias{stroma_classify}
\title{Stroma classifier}
\usage{
stroma_classify(input, broad.classify = TRUE, return.seurat = TRUE)
}
\arguments{
\item{input}{A Seurat object}

\item{broad.classify}{Whether to run "broad annotation" on hematopoietic cells (default: TRUE)}

\item{return.seurat}{Whether to return Seurat object, otherwise table (default: TRUE)}
}
\value{
A Seurat object with added annotations, or a table of classifier results
}
\description{
This performs annotation of mouse bone marrow stromal cell types using the
single cell references from Swann et al. (2024), and also optionally the
broad HemaScribe annotation on hematopoietic cells. Currently only supports
Seurat object as inputs.
}
