% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HemaScribe.R
\name{translate}
\alias{translate}
\title{Translate gene names to Ensembl IDs}
\usage{
translate(x)
}
\arguments{
\item{x}{List of gene names}
}
\value{
List of Ensembl IDs
}
\description{
Translate gene names to Ensembl IDs
}
