% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download reference data}
\usage{
download_data(dataDir = NA)
}
\arguments{
\item{dataDir}{Specifies where the reference data should be downloaded.  Should essentially never be altered by the user directly since the package looks for the data in a specific location.}
}
\description{
Downloads the reference data from the Github repo.  The user is prompted for
download when the package is first loaded, but this function can also be
called in a later session to update the reference data file.
}
