% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StromaScribe.R
\name{stroma_classify}
\alias{stroma_classify}
\title{Stroma classifier}
\usage{
stroma_classify(input, broad.classify = TRUE, return.full = FALSE)
}
\arguments{
\item{input}{A SingleCellExperiment or Seurat object}

\item{broad.classify}{Whether to run "broad annotation" on hematopoietic cells (default: TRUE)}

\item{return.full}{Whether to return full predictions directly (default: FALSE)}
}
\value{
An object of the same type as input with added annotations, or a dataframe with classifier results
}
\description{
This performs annotation of mouse bone marrow stromal cell types using the
single cell references from Swann et al. (2024), and also optionally the
broad HemaScribe annotation on hematopoietic cells.
}
